requirements.txt
# JNE Data Pipeline - Python Dependencies

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Environment
python-dotenv>=1.0.0

# Optional: For future Airflow integration
# apache-airflow>=2.7.0
